Adjust go111modules detection and fix lint (#180)
* Autodetect go version and use this to adjust GO111MODULES behaviour

Go 1.17+ has a changed default behaviour for GO111MODULES so autodetect if we are running go1.17+
and change the behaviour GO111MODULES to match this.

Signed-off-by: Andrew Thornton <art27@cantab.net>

* And remove GO_VERSION as it is now auto-calculated

The old GO_VERSION environment/configuration variable is poorly explained and appears to have
conflicting uses. Switch to the auto-calculated version.

Signed-off-by: Andrew Thornton <art27@cantab.net>

* Finally fix lint

There were multiple issues that shellcheck found. This commit fixes these.

Signed-off-by: Andrew Thornton <art27@cantab.net>

Signed-off-by: Andrew Thornton <art27@cantab.net>