call userprovided setup.sh and build.sh script if existing (#97)
* call userprovided setup.sh and build.sh script if existing

this allows to execute arbitrary command. first in a setup phase and
later in the loop over all targets.
the scripts are source to have access on all variables. XGOOS and
XGOARCH are set to the actual value for the execution time of the
script.

fixes: #96

* Adjust README.md

Signed-off-by: Andrew Thornton <art27@cantab.net>

* do not hardcode hooks directory but add option to xgo to mount a given directory into the container

setup.sh and build.sh are searched in that directory

Signed-off-by: Andrew Thornton <art27@cantab.net>
Co-authored-by: Andrew Thornton <art27@cantab.net>